+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Bug fixes and improvements
 
+- server: add metadata API to enable Apollo Federation for a table
+- console: add console support for Apollo Federation
+
 ## v2.10.0-beta.1
 
 ### Introducing Apollo Federation v1 support (experimental)